int32で桁溢れが発生しているためint系のデータ型をint64に統一

+++ deals.go
@@ -27,7 +27,7 @@ type DealsResponse struct {
 }
 
 type DealsResponseMeta struct {
-	TotalCount int32 `json:"total_count"`
+	TotalCount int64 `json:"total_count"`
 }
 
 type DealResponse struct {
@@ -36,7 +36,7 @@ type DealResponse struct {
 
 type GetDealOpts struct {
 	// 取引先ID
-	PartnerID int32 `url:"partner_id,omitempty"`
+	PartnerID int64 `url:"partner_id,omitempty"`
 	// 決済状況 (未決済: unsettled, 完了: settled)
 	Status string `url:"status,omitempty"`
 	// 収支区分 (収入: income, 支出: expense)
@@ -47,15 +47,15 @@ type GetDealOpts struct {
 	EndIssueDate string `url:"end_issue_date,omitempty"`
 	// 取引の債権債務行の表示（without: 表示しない(デフォルト), with: 表示する）
 	Accruals string `url:"accruals,omitempty"`
-	Offset   uint32 `url:"offset,omitempty"`
-	Limit    uint32 `url:"limit,omitempty"`
+	Offset   int64  `url:"offset,omitempty"`
+	Limit    int64  `url:"limit,omitempty"`
 }
 
 type Deal struct {
 	// 取引ID
-	ID uint64 `json:"id"`
+	ID int64 `json:"id"`
 	// 事業所ID
-	CompanyID int32 `json:"company_id"`
+	CompanyID int64 `json:"company_id"`
 	// 発生日 (yyyy-mm-dd)
 	IssueDate string `json:"issue_date"`
 	// 支払期日 (yyyy-mm-dd)
@@ -67,7 +67,7 @@ type Deal struct {
 	// 収支区分 (収入: income, 支出: expense)
 	Type *string `json:"type,omitempty"`
 	// 取引先ID
-	PartnerID int32 `json:"partner_id"`
+	PartnerID int64 `json:"partner_id"`
 	// 取引先コード
 	PartnerCode *string `json:"partner_code,omitempty"`
 	// 管理番号
@@ -85,36 +85,36 @@ type Deal struct {
 }
 
 type DealDetails struct {
-	ID uint64 `json:"id"`
+	ID int64 `json:"id"`
 	// 税区分コード
-	TaxCode int32 `json:"tax_code"`
+	TaxCode int64 `json:"tax_code"`
 	// 勘定科目ID
-	AccountItemID int32 `json:"account_item_id"`
+	AccountItemID int64 `json:"account_item_id"`
 	// 取引金額（税込で指定してください）
 	Amount int64 `json:"amount"`
 	// 品目ID
-	ItemID *int32 `json:"item_id,omitempty"`
+	ItemID *int64 `json:"item_id,omitempty"`
 	// 部門ID
-	SectionID *int32 `json:"section_id,omitempty"`
+	SectionID *int64 `json:"section_id,omitempty"`
 	// メモタグID
-	TagIDs *[]int32 `json:"tag_ids,omitempty"`
+	TagIDs *[]int64 `json:"tag_ids,omitempty"`
 	// セグメント１ID
-	Segment1TagID *int32 `json:"segment_1_tag_id,omitempty"`
+	Segment1TagID *int64 `json:"segment_1_tag_id,omitempty"`
 	// セグメント２ID
-	Segment2TagID *int32 `json:"segment_2_tag_id,omitempty"`
+	Segment2TagID *int64 `json:"segment_2_tag_id,omitempty"`
 	// セグメント３ID
-	Segment3TagID *int32 `json:"segment_3_tag_id,omitempty"`
+	Segment3TagID *int64 `json:"segment_3_tag_id,omitempty"`
 	// 備考
 	Description *string `json:"description,omitempty"`
 	// 消費税額（指定しない場合は自動で計算されます）
-	Vat *int32 `json:"vat,omitempty"`
+	Vat *int64 `json:"vat,omitempty"`
 	// 貸借（貸方: credit, 借方: debit）
 	EntrySide string `json:"entry_side"`
 }
 
 type DealRenews struct {
 	// +更新行ID
-	ID uint64 `json:"id"`
+	ID int64 `json:"id"`
 	// 更新日 (yyyy-mm-dd)
 	UpdateDate string `json:"update_date"`
 	// +更新の対象行ID
@@ -127,20 +127,20 @@ type DealRenews struct {
 
 type DealPayments struct {
 	// 取引行ID
-	ID uint64 `json:"id"`
+	ID int64 `json:"id"`
 	// 支払日
 	Date string `json:"date"`
 	// 口座区分 (銀行口座: bank_account, クレジットカード: credit_card, 現金: wallet, プライベート資金（法人の場合は役員借入金もしくは役員借入金、個人の場合は事業主貸もしくは事業主借）: private_account_item)
 	FromWalletableType string `json:"from_walletable_type,omitempty"`
 	// 口座ID（from_walletable_typeがprivate_account_itemの場合は勘定科目ID）
-	FromWalletableID int32 `json:"from_walletable_id,omitempty"`
+	FromWalletableID int64 `json:"from_walletable_id,omitempty"`
 	// 支払金額
 	Amount int64 `json:"amount"`
 }
 
 type DealReceipts struct {
 	// 証憑ID
-	ID int32 `json:"id"`
+	ID int64 `json:"id"`
 	// ステータス(unconfirmed:確認待ち、confirmed:確認済み、deleted:削除済み、ignored:無視)
 	Status string `json:"status"`
 	// メモ
@@ -160,7 +160,7 @@ type DealReceipts struct {
 
 type DealUser struct {
 	// ユーザーID
-	ID int32 `json:"id"`
+	ID int64 `json:"id"`
 	// メールアドレス
 	Email string `json:"email"`
 	// 表示名
@@ -173,11 +173,11 @@ type DealCreateParams struct {
 	// 収支区分 (収入: income, 支出: expense)
 	Type string `json:"type"`
 	// 事業所ID
-	CompanyID int32 `json:"company_id"`
+	CompanyID int64 `json:"company_id"`
 	// 支払期日(yyyy-mm-dd)
 	DueDate *string `json:"due_date,omitempty"`
 	// 取引先ID
-	PartnerID *int32 `json:"partner_id,omitempty"`
+	PartnerID *int64 `json:"partner_id,omitempty"`
 	// 取引先コード
 	PartnerCode *string `json:"partner_code,omitempty"`
 	// 管理番号
@@ -186,39 +186,39 @@ type DealCreateParams struct {
 	// 支払行一覧（配列）：未指定の場合、未決済の取引を作成します。
 	Payments *[]DealCreateParamsPayments `json:"payments,omitempty"`
 	// 証憑ファイルID（配列）
-	ReceiptIDs *[]int32 `json:"receipt_ids,omitempty"`
+	ReceiptIDs *[]int64 `json:"receipt_ids,omitempty"`
 }
 
 type DealCreateParamsDetails struct {
 	// 税区分コード
-	TaxCode int32 `json:"tax_code"`
+	TaxCode int64 `json:"tax_code"`
 	// 勘定科目ID
-	AccountItemID int32 `json:"account_item_id"`
+	AccountItemID int64 `json:"account_item_id"`
 	// 取引金額（税込で指定してください）
 	Amount int64 `json:"amount"`
 	// 品目ID
-	ItemID *int32 `json:"item_id,omitempty"`
+	ItemID *int64 `json:"item_id,omitempty"`
 	// 部門ID
-	SectionID *int32 `json:"section_id,omitempty"`
+	SectionID *int64 `json:"section_id,omitempty"`
 	// メモタグID
-	TagIDs *[]int32 `json:"tag_ids,omitempty"`
+	TagIDs *[]int64 `json:"tag_ids,omitempty"`
 	// セグメント１ID
-	Segment1TagID *int32 `json:"segment_1_tag_id,omitempty"`
+	Segment1TagID *int64 `json:"segment_1_tag_id,omitempty"`
 	// セグメント２ID
-	Segment2TagID *int32 `json:"segment_2_tag_id,omitempty"`
+	Segment2TagID *int64 `json:"segment_2_tag_id,omitempty"`
 	// セグメント３ID
-	Segment3TagID *int32 `json:"segment_3_tag_id,omitempty"`
+	Segment3TagID *int64 `json:"segment_3_tag_id,omitempty"`
 	// 備考
 	Description *string `json:"description,omitempty"`
 	// 消費税額（指定しない場合は自動で計算されます）
-	Vat *int32 `json:"vat,omitempty"`
+	Vat *int64 `json:"vat,omitempty"`
 }
 
 type DealCreateParamsPayments struct {
 	// 支払金額：payments指定時は必須
 	Amount int64 `json:"amount"`
 	// 口座ID（from_walletable_typeがprivate_account_itemの場合は勘定科目ID）：payments指定時は必須
-	FromWalletableID int32 `json:"from_walletable_id"`
+	FromWalletableID int64 `json:"from_walletable_id"`
 	// 口座区分 (銀行口座: bank_account, クレジットカード: credit_card, 現金: wallet, プライベート資金（法人の場合は役員借入金もしくは役員借入金、個人の場合は事業主貸もしくは事業主借）: private_account_item)：payments指定時は必須
 	FromWalletableType string `json:"from_walletable_type"`
 	// 支払日：payments指定時は必須
@@ -231,49 +231,49 @@ type DealUpdateParams struct {
 	// 収支区分 (収入: income, 支出: expense)
 	Type string `json:"type"`
 	// 事業所ID
-	CompanyID int32 `json:"company_id"`
+	CompanyID int64 `json:"company_id"`
 	// 支払期日(yyyy-mm-dd)
 	DueDate *string `json:"due_date,omitempty"`
 	// 取引先ID
-	PartnerID *int32 `json:"partner_id,omitempty"`
+	PartnerID *int64 `json:"partner_id,omitempty"`
 	// 取引先コード
 	PartnerCode *string `json:"partner_code,omitempty"`
 	// 管理番号
 	RefNumber *string                   `json:"ref_number,omitempty"`
 	Details   []DealUpdateParamsDetails `json:"details"`
 	// 証憑ファイルID（配列）
-	ReceiptIDs []int32 `json:"receipt_ids,omitempty"`
+	ReceiptIDs []int64 `json:"receipt_ids,omitempty"`
 }
 
 type DealUpdateParamsDetails struct {
 	// 取引行ID: 既存取引行を更新する場合に指定します。IDを指定しない取引行は、新規行として扱われ追加されます。また、detailsに含まれない既存の取引行は削除されます。更新後も残したい行は、必ず取引行IDを指定してdetailsに含めてください。
-	ID *uint64 `json:"id,omitempty"`
+	ID *int64 `json:"id,omitempty"`
 	// 税区分コード
-	TaxCode int32 `json:"tax_code"`
+	TaxCode int64 `json:"tax_code"`
 	// 勘定科目ID
-	AccountItemID int32 `json:"account_item_id"`
+	AccountItemID int64 `json:"account_item_id"`
 	// 取引金額（税込で指定してください）
 	Amount int64 `json:"amount"`
 	// 品目ID
-	ItemID *int32 `json:"item_id,omitempty"`
+	ItemID *int64 `json:"item_id,omitempty"`
 	// 部門ID
-	SectionID *int32 `json:"section_id,omitempty"`
+	SectionID *int64 `json:"section_id,omitempty"`
 	// メモタグID
-	TagIDs *[]int32 `json:"tag_ids,omitempty"`
+	TagIDs *[]int64 `json:"tag_ids,omitempty"`
 	// セグメント１ID
-	Segment1TagID *int32 `json:"segment_1_tag_id,omitempty"`
+	Segment1TagID *int64 `json:"segment_1_tag_id,omitempty"`
 	// セグメント２ID
-	Segment2TagID *int32 `json:"segment_2_tag_id,omitempty"`
+	Segment2TagID *int64 `json:"segment_2_tag_id,omitempty"`
 	// セグメント３ID
-	Segment3TagID *int32 `json:"segment_3_tag_id,omitempty"`
+	Segment3TagID *int64 `json:"segment_3_tag_id,omitempty"`
 	// 備考
 	Description *string `json:"description,omitempty"`
 	// 消費税額（指定しない場合は自動で計算されます）
-	Vat *int32 `json:"vat,omitempty"`
+	Vat *int64 `json:"vat,omitempty"`
 }
 
 func (c *Client) GetDeals(ctx context.Context, oauth2Token *oauth2.Token,
-	companyID uint32, opts GetDealOpts) (*DealsResponse, *oauth2.Token, error) {
+	companyID int64, opts GetDealOpts) (*DealsResponse, *oauth2.Token, error) {
 	var result DealsResponse
 
 	v, err := query.Values(opts)
@@ -291,7 +291,7 @@ func (c *Client) GetDeals(ctx context.Context, oauth2Token *oauth2.Token,
 
 func (c *Client) GetDeal(
 	ctx context.Context, oauth2Token *oauth2.Token,
-	companyID uint32, dealID uint64, opts GetDealOpts,
+	companyID int64, dealID int64, opts GetDealOpts,
 ) (*Deal, *oauth2.Token, error) {
 	var result DealResponse
 
@@ -321,7 +321,7 @@ func (c *Client) CreateDeal(
 
 func (c *Client) UpdateDeal(
 	ctx context.Context, oauth2Token *oauth2.Token,
-	dealID uint64, params DealUpdateParams,
+	dealID int64, params DealUpdateParams,
 ) (*Deal, *oauth2.Token, error) {
 	var result DealResponse
 	oauth2Token, err := c.call(ctx, path.Join(APIPathDeals, fmt.Sprint(dealID)), http.MethodPut, oauth2Token, nil, params, &result)
@@ -333,7 +333,7 @@ func (c *Client) UpdateDeal(
 
 func (c *Client) DestroyDeal(
 	ctx context.Context, oauth2Token *oauth2.Token,
-	companyID uint32, dealID uint64,
+	companyID int64, dealID int64,
 ) (*oauth2.Token, error) {
 	v, err := query.Values(nil)
 	if err != nil {
